update Auth component
