--- src/assets/styles/index.scss
@@ -1,133 +1,0 @@
-$color_black: #2d2a2a;
-$color_blue: #1380d7;
-
-#root {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-
-  .page {
-    flex: 1 1 auto;
-    overflow-y: auto;
-
-    .ui.container {
-      height: 100%;
-      padding: 30px 0;
-    }
-  }
-
-  //&::-webkit-scrollbar-track {
-  //  -webkit-box-shadow: inset 0 0 6px rgba(0,0,0,0.3);
-  //  border-radius: 10px;
-  //  background-color: #F5F5F5;
-  //}
-  //
-  //&::-webkit-scrollbar {
-  //  width: 12px;
-  //  background-color: #F5F5F5;
-  //}
-  //
-  //&::-webkit-scrollbar-thumb {
-  //  border-radius: 10px;
-  //  -webkit-box-shadow: inset 0 0 6px rgba(0,0,0,.3);
-  //  background-color: #D62929;
-  //}
-}
-
-.AppHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  min-height: 60px;
-  height: 60px;
-  padding: 0 50px;
-  background: $color_black;
-  color: #fff;
-
-  a {
-    color: #fff;
-
-    &:hover {
-      color: $color_blue;
-    }
-  }
-
-  .right_block {
-    a {
-      border-bottom: 1px solid #fff;
-      font-size: 16px;
-      &:hover {
-        border-bottom: 1px solid $color_blue;
-      }
-    }
-  }
-}
-
-//.ui.container {
-//
-//  &.AuthPage {
-//    display: flex;
-//    justify-content: center;
-//    align-items: center;
-//    flex-direction: column;
-//
-//    form {
-//      display: flex;
-//      flex-direction: column;
-//
-//      h2 {
-//        text-align: center;
-//      }
-//
-//      .ui.input {
-//        width: 350px;
-//        margin: 10px 0;
-//        input {
-//          text-align: center;
-//        }
-//      }
-//
-//      .ui.button {
-//        width: 150px;
-//        margin: 15px auto;
-//        color: #fff;
-//        background: $color_black;
-//      }
-//    }
-//  }
-//
-//  &.UserPage {
-//    display: flex;
-//    flex-direction: column;
-//
-//    .information {
-//      display: flex;
-//
-//      .ui.image {
-//        border-radius: 100%;
-//        border: 2px solid $color_black;
-//        background: #eee;
-//      }
-//
-//      .user_information {
-//        display: flex;
-//        flex-direction: column;
-//        justify-content: space-between;
-//        width: calc(100% - 300px);
-//        padding: 25px 50px;
-//
-//        .friends {
-//          margin-bottom: 50px;
-//        }
-//      }
-//    }
-//
-//    .PostsList {
-//      margin: 20px 0;
-//    }
-//  }
-//}
