+++ src/colors.scss
@@ -1,0 +1,2 @@
+$color_black: #2d2a2a;
+$color_blue: #1380d7;
