+++ src/components/Header/index.js
@@ -1,10 +1,11 @@
 import React from 'react';
 import {Link} from "react-router-dom";
-import {HOME_ROUTE, LOGIN_ROUTE} from "../utils/routesConstants";
+import {HOME_ROUTE, LOGIN_ROUTE} from "../../utils/routesConstants";
+import './styles.scss';
 
-const AppHeader = () => {
+const Header = () => {
     return (
-        <header className="AppHeader">
+        <header className="Header">
             <div className="left_block">
                 <h1><Link to={HOME_ROUTE} >SSN</Link></h1>
             </div>
@@ -15,4 +16,4 @@ const AppHeader = () => {
     );
 };
 
-export default AppHeader;
+export default Header;
