+++ src/index.scss
@@ -1,0 +1,21 @@
+@import "colors";
+
+#root {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+
+  .page {
+    flex: 1 1 auto;
+    overflow-y: auto;
+
+    .ui.container {
+      height: 100%;
+      padding: 30px 0;
+    }
+  }
+}
